Cleans up diff editor. Fixes #185944
