--- src/vs/editor/browser/widget/diffEditor.contribution.ts
@@ -1,124 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
-import { IDiffEditor } from 'vs/editor/browser/editorBrowser';
-import { ServicesAccessor } from 'vs/editor/browser/editorExtensions';
-import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
-import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
-import { localize } from 'vs/nls';
-import { ILocalizedString } from 'vs/platform/action/common/action';
-import { Action2, MenuId, MenuRegistry, registerAction2 } from 'vs/platform/actions/common/actions';
-import { CommandsRegistry } from 'vs/platform/commands/common/commands';
-import { ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
-import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
-
-const accessibleDiffViewerCategory: ILocalizedString = {
-	value: localize('accessibleDiffViewer', 'Accessible Diff Viewer'),
-	original: 'Accessible Diff Viewer',
-};
-
-export class AccessibleDiffViewerNext extends Action2 {
-	public static id = 'editor.action.accessibleDiffViewer.next';
-
-	constructor() {
-		super({
-			id: AccessibleDiffViewerNext.id,
-			title: { value: localize('editor.action.accessibleDiffViewer.next', "Go to Next Difference"), original: 'Go to Next Difference' },
-			category: accessibleDiffViewerCategory,
-			precondition: ContextKeyExpr.has('isInDiffEditor'),
-			keybinding: {
-				primary: KeyCode.F7,
-				weight: KeybindingWeight.EditorContrib
-			},
-			f1: true,
-		});
-	}
-
-	public override run(accessor: ServicesAccessor): void {
-		const diffEditor = findFocusedDiffEditor(accessor);
-		diffEditor?.accessibleDiffViewerNext();
-	}
-}
-
-MenuRegistry.appendMenuItem(MenuId.EditorTitle, {
-	command: {
-		id: AccessibleDiffViewerNext.id,
-		title: localize('Open Accessible Diff Viewer', "Open Accessible Diff Viewer"),
-	},
-	order: 10,
-	group: '2_diff',
-	when: ContextKeyExpr.and(
-		EditorContextKeys.accessibleDiffViewerVisible.negate(),
-		ContextKeyExpr.has('isInDiffEditor'),
-	),
-});
-
-export class AccessibleDiffViewerPrev extends Action2 {
-	public static id = 'editor.action.accessibleDiffViewer.prev';
-
-	constructor() {
-		super({
-			id: AccessibleDiffViewerPrev.id,
-			title: { value: localize('editor.action.accessibleDiffViewer.prev', "Go to Previous Difference"), original: 'Go to Previous Difference' },
-			category: accessibleDiffViewerCategory,
-			precondition: ContextKeyExpr.has('isInDiffEditor'),
-			keybinding: {
-				primary: KeyMod.Shift | KeyCode.F7,
-				weight: KeybindingWeight.EditorContrib
-			},
-			f1: true,
-		});
-	}
-
-	public override run(accessor: ServicesAccessor): void {
-		const diffEditor = findFocusedDiffEditor(accessor);
-		diffEditor?.accessibleDiffViewerPrev();
-	}
-}
-
-export function findFocusedDiffEditor(accessor: ServicesAccessor): IDiffEditor | null {
-	const codeEditorService = accessor.get(ICodeEditorService);
-	const diffEditors = codeEditorService.listDiffEditors();
-	const activeCodeEditor = codeEditorService.getFocusedCodeEditor() ?? codeEditorService.getActiveCodeEditor();
-	if (!activeCodeEditor) {
-		return null;
-	}
-
-	for (let i = 0, len = diffEditors.length; i < len; i++) {
-		const diffEditor = <IDiffEditor>diffEditors[i];
-		if (diffEditor.getModifiedEditor().getId() === activeCodeEditor.getId() || diffEditor.getOriginalEditor().getId() === activeCodeEditor.getId()) {
-			return diffEditor;
-		}
-	}
-
-	if (document.activeElement) {
-		for (const d of diffEditors) {
-			const container = d.getContainerDomNode();
-			if (isElementOrParentOf(container, document.activeElement)) {
-				return d;
-			}
-		}
-	}
-
-	return null;
-}
-
-function isElementOrParentOf(elementOrParent: Element, element: Element): boolean {
-	let e: Element | null = element;
-	while (e) {
-		if (e === elementOrParent) {
-			return true;
-		}
-		e = e.parentElement;
-	}
-	return false;
-}
-
-CommandsRegistry.registerCommandAlias('editor.action.diffReview.next', AccessibleDiffViewerNext.id);
-registerAction2(AccessibleDiffViewerNext);
-
-CommandsRegistry.registerCommandAlias('editor.action.diffReview.prev', AccessibleDiffViewerPrev.id);
-registerAction2(AccessibleDiffViewerPrev);
