+++ src/vs/editor/browser/widget/diffEditor/accessibleDiffViewer.css
