+++ src/vs/editor/browser/widget/diffEditor/colors.ts
