+++ src/vs/editor/browser/widget/diffEditor/decorations.ts
