+++ src/vs/editor/browser/widget/diffEditor/delegatingEditorImpl.ts
