+++ src/vs/editor/browser/widget/diffEditor/diffEditor.contribution.ts
@@ -4,17 +4,19 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Codicon } from 'vs/base/common/codicons';
-import { KeyCode } from 'vs/base/common/keyCodes';
-import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
+import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
+import { ICodeEditor, IDiffEditor } from 'vs/editor/browser/editorBrowser';
 import { EditorAction2, ServicesAccessor } from 'vs/editor/browser/editorExtensions';
-import { findFocusedDiffEditor } from 'vs/editor/browser/widget/diffEditor.contribution';
-import { DiffEditorWidget2 } from 'vs/editor/browser/widget/diffEditorWidget2/diffEditorWidget2';
+import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
+import { DiffEditorWidget } from 'vs/editor/browser/widget/diffEditor/diffEditorWidget';
 import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
 import { localize } from 'vs/nls';
 import { ILocalizedString } from 'vs/platform/action/common/action';
 import { Action2, MenuId, MenuRegistry, registerAction2 } from 'vs/platform/actions/common/actions';
+import { CommandsRegistry } from 'vs/platform/commands/common/commands';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { ContextKeyEqualsExpr, ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
+import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
 
 export class ToggleCollapseUnchangedRegions extends Action2 {
 	constructor() {
@@ -22,13 +24,11 @@ export class ToggleCollapseUnchangedRegions extends Action2 {
 			id: 'diffEditor.toggleCollapseUnchangedRegions',
 			title: { value: localize('toggleCollapseUnchangedRegions', "Toggle Collapse Unchanged Regions"), original: 'Toggle Collapse Unchanged Regions' },
 			icon: Codicon.map,
-			precondition: ContextKeyEqualsExpr.create('diffEditorVersion', 2),
 			toggled: ContextKeyExpr.has('config.diffEditor.hideUnchangedRegions.enabled'),
 			menu: {
 				id: MenuId.EditorTitle,
 				order: 22,
 				group: 'navigation',
-				when: ContextKeyEqualsExpr.create('diffEditorVersion', 2),
 			},
 		});
 	}
@@ -47,7 +47,6 @@ export class ToggleShowMovedCodeBlocks extends Action2 {
 		super({
 			id: 'diffEditor.toggleShowMovedCodeBlocks',
 			title: { value: localize('toggleShowMovedCodeBlocks', "Toggle Show Moved Code Blocks"), original: 'Toggle Show Moved Code Blocks' },
-			precondition: ContextKeyEqualsExpr.create('diffEditorVersion', 2),
 		});
 	}
 
@@ -65,7 +64,6 @@ export class ToggleUseInlineViewWhenSpaceIsLimited extends Action2 {
 		super({
 			id: 'diffEditor.toggleUseInlineViewWhenSpaceIsLimited',
 			title: { value: localize('toggleUseInlineViewWhenSpaceIsLimited', "Toggle Use Inline View When Space Is Limited"), original: 'Toggle Use Inline View When Space Is Limited' },
-			precondition: ContextKeyEqualsExpr.create('diffEditorVersion', 2),
 		});
 	}
 
@@ -86,10 +84,7 @@ MenuRegistry.appendMenuItem(MenuId.EditorTitle, {
 	},
 	order: 11,
 	group: '1_diff',
-	when: ContextKeyExpr.and(
-		EditorContextKeys.diffEditorRenderSideBySideInlineBreakpointReached,
-		ContextKeyEqualsExpr.create('diffEditorVersion', 2)
-	)
+	when: EditorContextKeys.diffEditorRenderSideBySideInlineBreakpointReached,
 });
 
 MenuRegistry.appendMenuItem(MenuId.EditorTitle, {
@@ -101,7 +96,6 @@ MenuRegistry.appendMenuItem(MenuId.EditorTitle, {
 	},
 	order: 10,
 	group: '1_diff',
-	when: ContextKeyEqualsExpr.create('diffEditorVersion', 2)
 });
 
 const diffEditorCategory: ILocalizedString = {
@@ -115,7 +109,7 @@ export class SwitchSide extends EditorAction2 {
 			id: 'diffEditor.switchSide',
 			title: { value: localize('switchSide', "Switch Side"), original: 'Switch Side' },
 			icon: Codicon.arrowSwap,
-			precondition: ContextKeyExpr.and(ContextKeyEqualsExpr.create('diffEditorVersion', 2), ContextKeyExpr.has('isInDiffEditor')),
+			precondition: ContextKeyExpr.has('isInDiffEditor'),
 			f1: true,
 			category: diffEditorCategory,
 		});
@@ -123,7 +117,7 @@ export class SwitchSide extends EditorAction2 {
 
 	runEditorCommand(accessor: ServicesAccessor, editor: ICodeEditor, arg?: { dryRun: boolean }): unknown {
 		const diffEditor = findFocusedDiffEditor(accessor);
-		if (diffEditor instanceof DiffEditorWidget2) {
+		if (diffEditor instanceof DiffEditorWidget) {
 			if (arg && arg.dryRun) {
 				return { destinationSelection: diffEditor.mapToOtherSide().destinationSelection };
 			} else {
@@ -154,7 +148,7 @@ export class ExitCompareMove extends EditorAction2 {
 
 	runEditorCommand(accessor: ServicesAccessor, editor: ICodeEditor, ...args: unknown[]): void {
 		const diffEditor = findFocusedDiffEditor(accessor);
-		if (diffEditor instanceof DiffEditorWidget2) {
+		if (diffEditor instanceof DiffEditorWidget) {
 			diffEditor.exitCompareMove();
 		}
 	}
@@ -168,7 +162,7 @@ export class CollapseAllUnchangedRegions extends EditorAction2 {
 			id: 'diffEditor.collapseAllUnchangedRegions',
 			title: { value: localize('collapseAllUnchangedRegions', "Collapse All Unchanged Regions"), original: 'Collapse All Unchanged Regions' },
 			icon: Codicon.fold,
-			precondition: ContextKeyExpr.and(ContextKeyEqualsExpr.create('diffEditorVersion', 2), ContextKeyExpr.has('isInDiffEditor')),
+			precondition: ContextKeyExpr.has('isInDiffEditor'),
 			f1: true,
 			category: diffEditorCategory,
 		});
@@ -176,7 +170,7 @@ export class CollapseAllUnchangedRegions extends EditorAction2 {
 
 	runEditorCommand(accessor: ServicesAccessor, editor: ICodeEditor, ...args: unknown[]): void {
 		const diffEditor = findFocusedDiffEditor(accessor);
-		if (diffEditor instanceof DiffEditorWidget2) {
+		if (diffEditor instanceof DiffEditorWidget) {
 			diffEditor.collapseAllUnchangedRegions();
 		}
 	}
@@ -190,7 +184,7 @@ export class ShowAllUnchangedRegions extends EditorAction2 {
 			id: 'diffEditor.showAllUnchangedRegions',
 			title: { value: localize('showAllUnchangedRegions', "Show All Unchanged Regions"), original: 'Show All Unchanged Regions' },
 			icon: Codicon.unfold,
-			precondition: ContextKeyExpr.and(ContextKeyEqualsExpr.create('diffEditorVersion', 2), ContextKeyExpr.has('isInDiffEditor')),
+			precondition: ContextKeyExpr.has('isInDiffEditor'),
 			f1: true,
 			category: diffEditorCategory,
 		});
@@ -198,10 +192,118 @@ export class ShowAllUnchangedRegions extends EditorAction2 {
 
 	runEditorCommand(accessor: ServicesAccessor, editor: ICodeEditor, ...args: unknown[]): void {
 		const diffEditor = findFocusedDiffEditor(accessor);
-		if (diffEditor instanceof DiffEditorWidget2) {
+		if (diffEditor instanceof DiffEditorWidget) {
 			diffEditor.showAllUnchangedRegions();
 		}
 	}
 }
 
 registerAction2(ShowAllUnchangedRegions);
+
+const accessibleDiffViewerCategory: ILocalizedString = {
+	value: localize('accessibleDiffViewer', 'Accessible Diff Viewer'),
+	original: 'Accessible Diff Viewer',
+};
+
+export class AccessibleDiffViewerNext extends Action2 {
+	public static id = 'editor.action.accessibleDiffViewer.next';
+
+	constructor() {
+		super({
+			id: AccessibleDiffViewerNext.id,
+			title: { value: localize('editor.action.accessibleDiffViewer.next', "Go to Next Difference"), original: 'Go to Next Difference' },
+			category: accessibleDiffViewerCategory,
+			precondition: ContextKeyExpr.has('isInDiffEditor'),
+			keybinding: {
+				primary: KeyCode.F7,
+				weight: KeybindingWeight.EditorContrib
+			},
+			f1: true,
+		});
+	}
+
+	public override run(accessor: ServicesAccessor): void {
+		const diffEditor = findFocusedDiffEditor(accessor);
+		diffEditor?.accessibleDiffViewerNext();
+	}
+}
+
+MenuRegistry.appendMenuItem(MenuId.EditorTitle, {
+	command: {
+		id: AccessibleDiffViewerNext.id,
+		title: localize('Open Accessible Diff Viewer', "Open Accessible Diff Viewer"),
+	},
+	order: 10,
+	group: '2_diff',
+	when: ContextKeyExpr.and(
+		EditorContextKeys.accessibleDiffViewerVisible.negate(),
+		ContextKeyExpr.has('isInDiffEditor'),
+	),
+});
+
+export class AccessibleDiffViewerPrev extends Action2 {
+	public static id = 'editor.action.accessibleDiffViewer.prev';
+
+	constructor() {
+		super({
+			id: AccessibleDiffViewerPrev.id,
+			title: { value: localize('editor.action.accessibleDiffViewer.prev', "Go to Previous Difference"), original: 'Go to Previous Difference' },
+			category: accessibleDiffViewerCategory,
+			precondition: ContextKeyExpr.has('isInDiffEditor'),
+			keybinding: {
+				primary: KeyMod.Shift | KeyCode.F7,
+				weight: KeybindingWeight.EditorContrib
+			},
+			f1: true,
+		});
+	}
+
+	public override run(accessor: ServicesAccessor): void {
+		const diffEditor = findFocusedDiffEditor(accessor);
+		diffEditor?.accessibleDiffViewerPrev();
+	}
+}
+
+export function findFocusedDiffEditor(accessor: ServicesAccessor): IDiffEditor | null {
+	const codeEditorService = accessor.get(ICodeEditorService);
+	const diffEditors = codeEditorService.listDiffEditors();
+	const activeCodeEditor = codeEditorService.getFocusedCodeEditor() ?? codeEditorService.getActiveCodeEditor();
+	if (!activeCodeEditor) {
+		return null;
+	}
+
+	for (let i = 0, len = diffEditors.length; i < len; i++) {
+		const diffEditor = <IDiffEditor>diffEditors[i];
+		if (diffEditor.getModifiedEditor().getId() === activeCodeEditor.getId() || diffEditor.getOriginalEditor().getId() === activeCodeEditor.getId()) {
+			return diffEditor;
+		}
+	}
+
+	if (document.activeElement) {
+		for (const d of diffEditors) {
+			const container = d.getContainerDomNode();
+			if (isElementOrParentOf(container, document.activeElement)) {
+				return d;
+			}
+		}
+	}
+
+	return null;
+}
+
+function isElementOrParentOf(elementOrParent: Element, element: Element): boolean {
+	let e: Element | null = element;
+	while (e) {
+		if (e === elementOrParent) {
+			return true;
+		}
+		e = e.parentElement;
+	}
+	return false;
+}
+
+CommandsRegistry.registerCommandAlias('editor.action.diffReview.next', AccessibleDiffViewerNext.id);
+registerAction2(AccessibleDiffViewerNext);
+
+CommandsRegistry.registerCommandAlias('editor.action.diffReview.prev', AccessibleDiffViewerPrev.id);
+registerAction2(AccessibleDiffViewerPrev);
