+++ src/vs/editor/browser/widget/diffEditor/diffEditorSash.ts
