+++ src/vs/editor/browser/widget/diffEditor/diffEditorViewModel.ts
@@ -8,8 +8,8 @@ import { CancellationTokenSource } from 'vs/base/common/cancellation';
 import { Disposable, toDisposable } from 'vs/base/common/lifecycle';
 import { IObservable, IReader, ISettableObservable, ITransaction, autorunWithStore, derived, observableSignal, observableSignalFromEvent, observableValue, transaction, waitForState } from 'vs/base/common/observable';
 import { IDiffEditor } from 'vs/editor/browser/editorBrowser';
-import { IDiffProviderFactoryService } from 'vs/editor/browser/widget/diffEditorWidget2/diffProviderFactoryService';
-import { readHotReloadableExport } from 'vs/editor/browser/widget/diffEditorWidget2/utils';
+import { IDiffProviderFactoryService } from 'vs/editor/browser/widget/diffEditor/diffProviderFactoryService';
+import { readHotReloadableExport } from 'vs/editor/browser/widget/diffEditor/utils';
 import { ISerializedLineRange, LineRange } from 'vs/editor/common/core/lineRange';
 import { DefaultLinesDiffComputer } from 'vs/editor/common/diff/defaultLinesDiffComputer/defaultLinesDiffComputer';
 import { IDocumentDiff } from 'vs/editor/common/diff/documentDiffProvider';
