+++ src/vs/editor/browser/widget/diffEditor/outlineModel.ts
