+++ src/vs/editor/browser/widget/diffEditor/renderLines.ts
