+++ src/vs/editor/browser/widget/diffEditor/utils.ts
