+++ src/vs/editor/browser/widget/diffEditor/workerBasedDocumentDiffProvider.ts
