--- src/vs/editor/browser/widget/diffEditorWidget2/style.css
@@ -1,151 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-.monaco-editor .diff-hidden-lines-widget {
-	width: 100%;
-}
-
-.monaco-editor .diff-hidden-lines {
-	height: 0px; /* The children each have a fixed height, the transform confuses the browser */
-	transform: translate(0px, -10px);
-	font-size: 13px;
-	line-height: 14px;
-}
-
-.monaco-editor .diff-hidden-lines:not(.dragging) .top:hover,
-.monaco-editor .diff-hidden-lines:not(.dragging) .bottom:hover,
-.monaco-editor .diff-hidden-lines .top.dragging,
-.monaco-editor .diff-hidden-lines .bottom.dragging {
-	background-color: var(--vscode-focusBorder);
-}
-
-.monaco-editor .diff-hidden-lines .top,
-.monaco-editor .diff-hidden-lines .bottom {
-	transition: background-color 0.1s ease-out;
-	height: 4px;
-	background-color: transparent;
-	background-clip: padding-box;
-	border-bottom: 2px solid transparent;
-	border-top: 4px solid transparent;
-	cursor: ns-resize;
-}
-
-.monaco-editor .diff-hidden-lines .top {
-	transform: translate(0px, 4px);
-}
-
-.monaco-editor .diff-hidden-lines .bottom {
-	transform: translate(0px, -6px);
-}
-
-.monaco-editor .diff-unchanged-lines {
-	background: var(--vscode-diffEditor-unchangedCodeBackground);
-}
-
-.monaco-editor .noModificationsOverlay {
-	z-index: 1;
-	background: var(--vscode-editor-background);
-
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-
-
-.monaco-editor .diff-hidden-lines .center {
-	background: var(--vscode-diffEditor-unchangedRegionBackground);
-	color: var(--vscode-diffEditor-unchangedRegionForeground);
-	overflow: hidden;
-	display: block;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-
-	height: 24px;
-}
-
-.monaco-editor .diff-hidden-lines .center span.codicon {
-	vertical-align: middle;
-}
-
-.monaco-editor .diff-hidden-lines .center a:hover .codicon {
-	cursor: pointer;
-	color: var(--vscode-editorLink-activeForeground) !important;
-}
-
-.monaco-editor .diff-hidden-lines div.breadcrumb-item {
-	cursor: pointer;
-}
-
-.monaco-editor .diff-hidden-lines div.breadcrumb-item:hover {
-	color: var(--vscode-editorLink-activeForeground);
-}
-
-.monaco-editor .movedOriginal {
-	border: 2px solid var(--vscode-diffEditor-move-border);
-}
-
-.monaco-editor .movedModified {
-	border: 2px solid var(--vscode-diffEditor-move-border);
-}
-
-.monaco-editor .movedOriginal.currentMove, .monaco-editor .movedModified.currentMove {
-	border: 2px solid var(--vscode-diffEditor-moveActive-border);
-}
-
-.monaco-diff-editor .moved-blocks-lines path.currentMove {
-	stroke: var(--vscode-diffEditor-moveActive-border);
-}
-
-.monaco-diff-editor .moved-blocks-lines path {
-	pointer-events: visiblestroke;
-}
-
-.monaco-diff-editor .moved-blocks-lines .arrow {
-	fill: var(--vscode-diffEditor-move-border);
-}
-
-.monaco-diff-editor .moved-blocks-lines .arrow.currentMove {
-	fill: var(--vscode-diffEditor-moveActive-border);
-}
-
-.monaco-diff-editor .moved-blocks-lines .arrow-rectangle {
-	fill: var(--vscode-editor-background);
-}
-
-.monaco-diff-editor .moved-blocks-lines {
-	position: absolute;
-	pointer-events: none;
-}
-
-.monaco-diff-editor .moved-blocks-lines path {
-	fill: none;
-	stroke: var(--vscode-diffEditor-move-border);
-	stroke-width: 2;
-}
-
-.monaco-editor .char-delete.diff-range-empty {
-	margin-left: -1px;
-	border-left: solid var(--vscode-diffEditor-removedTextBackground) 3px;
-}
-
-.monaco-editor .char-insert.diff-range-empty {
-	border-left: solid var(--vscode-diffEditor-insertedTextBackground) 3px;
-}
-
-.monaco-editor .fold-unchanged {
-	cursor: pointer;
-}
-
-.monaco-diff-editor .diff-moved-code-block {
-	display: flex;
-	justify-content: flex-end;
-	margin-top: -4px;
-}
-
-.monaco-diff-editor .diff-moved-code-block .action-bar .action-label.codicon {
-	width: 12px;
-	height: 12px;
-	font-size: 12px;
-}
