+++ src/vs/editor/test/browser/widget/diffEditorWidget2.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import { UnchangedRegion } from 'vs/editor/browser/widget/diffEditorWidget2/diffEditorViewModel';
+import { UnchangedRegion } from 'vs/editor/browser/widget/diffEditor/diffEditorViewModel';
 import { LineRange } from 'vs/editor/common/core/lineRange';
 import { DetailedLineRangeMapping } from 'vs/editor/common/diff/rangeMapping';
